@@ -13,15 +13,15 @@ dependencies:
   
   flutter:
     sdk: flutter
-  path: ^1.8.3
-  path_provider: ^2.1.1
-  powersync: ^0.4.1
+  path: ">=1.8.3 <1.9.0"
+  path_provider: ^2.1.2
+  powersync: ^1.1.0
   shared:
     path: ../shared
-  supabase_flutter: ^1.10.22
+  supabase_flutter: ^2.2.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mocktail: ^1.0.0
+  mocktail: ^1.0.2
   very_good_analysis: ^5.1.0
