@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use
+
 import 'dart:async';
 import 'dart:io';
 
@@ -14,7 +16,7 @@ import 'package:stories_editor/src/presentation/widgets/color_selector.dart';
 import 'package:stories_editor/src/presentation/widgets/size_slider_selector.dart';
 
 class Painting extends StatefulWidget {
-  const Painting({Key? key}) : super(key: key);
+  const Painting({super.key});
 
   @override
   State<Painting> createState() => _PaintingState();
@@ -42,8 +44,8 @@ class _PaintingState extends State<Painting> {
     var screenSize = MediaQueryData.fromView(View.of(context));
 
     /// on gestures start
-    void _onPanStart(DragStartDetails details,
-        PaintingNotifier paintingNotifier, ControlNotifier controlProvider) {
+    void onPanStart(DragStartDetails details, PaintingNotifier paintingNotifier,
+        ControlNotifier controlProvider) {
       final box = context.findRenderObject() as RenderBox;
       final offset = box.globalToLocal(details.globalPosition);
       final point = Point(offset.dx, offset.dy);
@@ -69,7 +71,7 @@ class _PaintingState extends State<Painting> {
     }
 
     /// on gestures update
-    void _onPanUpdate(DragUpdateDetails details,
+    void onPanUpdate(DragUpdateDetails details,
         PaintingNotifier paintingNotifier, ControlNotifier controlNotifier) {
       final box = context.findRenderObject() as RenderBox;
       final offset = box.globalToLocal(details.globalPosition);
@@ -97,24 +99,24 @@ class _PaintingState extends State<Painting> {
     }
 
     /// on gestures end
-    void _onPanEnd(DragEndDetails details, PaintingNotifier paintingNotifier) {
+    void onPanEnd(DragEndDetails details, PaintingNotifier paintingNotifier) {
       paintingNotifier.lines = List.from(paintingNotifier.lines)..add(line!);
       line = null;
       paintingNotifier.linesStreamController.add(paintingNotifier.lines);
     }
 
     /// paint current line
-    Widget _renderCurrentLine(BuildContext context,
+    Widget renderCurrentLine(BuildContext context,
         PaintingNotifier paintingNotifier, ControlNotifier controlNotifier) {
       return GestureDetector(
         onPanStart: (details) {
-          _onPanStart(details, paintingNotifier, controlNotifier);
+          onPanStart(details, paintingNotifier, controlNotifier);
         },
         onPanUpdate: (details) {
-          _onPanUpdate(details, paintingNotifier, controlNotifier);
+          onPanUpdate(details, paintingNotifier, controlNotifier);
         },
         onPanEnd: (details) {
-          _onPanEnd(details, paintingNotifier);
+          onPanEnd(details, paintingNotifier);
         },
         child: RepaintBoundary(
           child: SafeArea(
@@ -161,7 +163,7 @@ class _PaintingState extends State<Painting> {
             body: Stack(
               children: [
                 /// render current line
-                _renderCurrentLine(context, paintingNotifier, controlNotifier),
+                renderCurrentLine(context, paintingNotifier, controlNotifier),
 
                 /// select line width
                 const Align(
