@@ -8,16 +8,16 @@ import 'package:stories_editor/src/presentation/utils/constants/app_enums.dart';
 import 'package:stories_editor/src/presentation/widgets/animated_onTap_button.dart';
 
 class AnimationSelector extends StatelessWidget {
-  const AnimationSelector({Key? key}) : super(key: key);
+  const AnimationSelector({super.key});
 
   @override
   Widget build(BuildContext context) {
-    var _size = MediaQuery.of(context).size;
+    var size = MediaQuery.of(context).size;
     return Consumer2<TextEditingNotifier, ControlNotifier>(
       builder: (context, editorNotifier, controlNotifier, child) {
         return Container(
-          height: _size.width * 0.1,
-          width: _size.width,
+          height: size.width * 0.1,
+          width: size.width,
           alignment: Alignment.center,
           child: PageView.builder(
             controller: editorNotifier.textAnimationController,
@@ -36,8 +36,8 @@ class AnimationSelector extends StatelessWidget {
                   editorNotifier.textAnimationController.jumpToPage(index);
                 },
                 child: Container(
-                  height: _size.width * 0.25,
-                  width: _size.width * 0.25,
+                  height: size.width * 0.25,
+                  width: size.width * 0.25,
                   alignment: Alignment.center,
                   margin: const EdgeInsets.all(2),
                   decoration: BoxDecoration(
