@@ -7,12 +7,12 @@ import 'package:stories_editor/src/domain/providers/notifiers/control_provider.d
 import 'package:stories_editor/src/domain/providers/notifiers/text_editing_notifier.dart';
 
 class TextFieldWidget extends StatelessWidget {
-  const TextFieldWidget({Key? key}) : super(key: key);
+  const TextFieldWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
     final ScreenUtil screenUtil = ScreenUtil();
-    FocusNode _textNode = FocusNode();
+    FocusNode textNode = FocusNode();
     return Consumer2<TextEditingNotifier, ControlNotifier>(
       builder: (context, editorNotifier, controlNotifier, child) {
         return Center(
@@ -30,14 +30,14 @@ class TextFieldWidget extends StatelessWidget {
                   padding: const EdgeInsets.only(right: 2),
                   child: _text(
                     editorNotifier: editorNotifier,
-                    textNode: _textNode,
+                    textNode: textNode,
                     controlNotifier: controlNotifier,
                     paintingStyle: PaintingStyle.fill,
                   ),
                 ),
                 _textField(
                   editorNotifier: editorNotifier,
-                  textNode: _textNode,
+                  textNode: textNode,
                   controlNotifier: controlNotifier,
                   paintingStyle: PaintingStyle.stroke,
                 )
