@@ -8,15 +8,15 @@ class ToolButton extends StatelessWidget {
   final EdgeInsets? padding;
   final Function()? onLongPress;
   final Color colorBorder;
-  const ToolButton(
-      {Key? key,
-      required this.onTap,
-      required this.child,
-      this.backGroundColor,
-      this.padding,
-      this.onLongPress,
-      this.colorBorder = Colors.white})
-      : super(key: key);
+  const ToolButton({
+    super.key,
+    required this.onTap,
+    required this.child,
+    this.backGroundColor,
+    this.padding,
+    this.onLongPress,
+    this.colorBorder = Colors.white,
+  });
 
   @override
   Widget build(BuildContext context) {
