@@ -1,5 +1,5 @@
+export '../../stories/widgets/stories_carousel.dart';
 export 'divider_block.dart';
 export 'feed_app_bar.dart';
 export 'feed_item_loader.dart';
 export 'feed_page_controller.dart';
-export '../../stories/widgets/stories_carousel.dart';
