@@ -26,7 +26,7 @@ class NetworkError extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final l10n = context.l10n;
-    final theme = Theme.of(context);
+    final theme = context.theme;
     return Column(
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
