@@ -3,7 +3,8 @@ import 'package:flutter/material.dart' show Color;
 /// Converts from a hex [String] to [Color]. Converts a [Color] to a hex
 /// [String].
 extension HexColor on Color {
-  /// String is in the format "aabbcc" or "ffaabbcc" with an optional leading "#".
+  /// String is in the format "aabbcc" or "ffaabbcc" with an optional leading 
+  /// "#".
   static Color fromHex(String hexString) {
     final buffer = StringBuffer();
     if (hexString.length == 6 || hexString.length == 7) buffer.write('ff');
@@ -11,7 +12,8 @@ extension HexColor on Color {
     return Color(int.parse(buffer.toString(), radix: 16));
   }
 
-  /// Prefixes a hash sign if [leadingHashSign] is set to `true` (default is `true`).
+  /// Prefixes a hash sign if [leadingHashSign] is set to `true` (default is 
+  /// `true`).
   String toHex({bool leadingHashSign = true}) => '${leadingHashSign ? '#' : ''}'
       '${alpha.toRadixString(16).padLeft(2, '0')}'
       '${red.toRadixString(16).padLeft(2, '0')}'
