@@ -35,7 +35,7 @@ extension BlocActionX on BlockAction {
       };
 }
 
-/// {@macro block_action}
+/// {@template block_action}
 /// A class which represents an action that can occur
 /// when interacting with a block.
 /// {@endtemplate}
