@@ -1,4 +1,5 @@
-/// A package that manages connection to the Powersync cloud service and database.
+/// A package that manages connection to the Powersync cloud service and 
+/// database.
 library powersync_repository;
 
 export 'package:supabase_flutter/supabase_flutter.dart' hide User;
