@@ -4,6 +4,6 @@ export 'debouncer.dart';
 export 'image_picker.dart';
 export 'logger.dart';
 export 'throttler.dart';
+export 'tld.dart';
 export 'uid.dart';
 export 'utilities/utilities.dart';
-export 'tld.dart';
