@@ -1,3 +1,5 @@
+// ignore_for_file: public_member_api_docs
+
 import 'dart:io';
 
 import 'package:app_ui/app_ui.dart';
@@ -7,7 +9,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_image_compress/flutter_image_compress.dart';
 import 'package:insta_assets_picker/insta_assets_picker.dart';
 import 'package:shared/shared.dart';
-import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 
 class PickImage {
   /// {@macro image_picker}
