@@ -1,4 +1,4 @@
-/// A Secure Storage client which implements the base [Storage] interface.
+/// A Secure Storage client which implements the base `Storage` interface.
 library secure_storage;
 
 export 'src/secure_storage.dart';
