@@ -1,31 +1,42 @@
-/// Generated file. Do not edit.
-
 // ignore_for_file: constant_identifier_names
-// ignore_for_file: lines_longer_than_80_chars
 class PackagesRes {
   static const String PLUGIN_NAME = 'stories_editor';
   static const String PLUGIN_VERSION = '0.2.2';
-  static const String ALIGN_CENTER = 'packages/stories_editor/assets/icons/align_center.png';
-  static const String ALIGN_LEFT = 'packages/stories_editor/assets/icons/align_left.png';
-  static const String ALIGN_RIGTH = 'packages/stories_editor/assets/icons/align_rigth.png';
+  static const String ALIGN_CENTER =
+      'packages/stories_editor/assets/icons/align_center.png';
+  static const String ALIGN_LEFT =
+      'packages/stories_editor/assets/icons/align_left.png';
+  static const String ALIGN_RIGTH =
+      'packages/stories_editor/assets/icons/align_rigth.png';
   static const String ARROW = 'packages/stories_editor/assets/icons/arrow.png';
   static const String BRUSH = 'packages/stories_editor/assets/icons/brush.png';
   static const String CHECK = 'packages/stories_editor/assets/icons/check.png';
-  static const String CIRCULAR_GRADIENT = 'packages/stories_editor/assets/icons/circular_gradient.png';
-  static const String COLORPALETTE = 'packages/stories_editor/assets/icons/colorpalette.png';
-  static const String DOWNLOAD = 'packages/stories_editor/assets/icons/download.png';
+  static const String CIRCULAR_GRADIENT =
+      'packages/stories_editor/assets/icons/circular_gradient.png';
+  static const String COLORPALETTE =
+      'packages/stories_editor/assets/icons/colorpalette.png';
+  static const String DOWNLOAD =
+      'packages/stories_editor/assets/icons/download.png';
   static const String DRAFT = 'packages/stories_editor/assets/icons/draft.png';
   static const String DRAW = 'packages/stories_editor/assets/icons/draw.png';
-  static const String FONT_BACKGROUND = 'packages/stories_editor/assets/icons/font_backGround.png';
-  static const String MARKER = 'packages/stories_editor/assets/icons/marker.png';
+  static const String FONT_BACKGROUND =
+      'packages/stories_editor/assets/icons/font_backGround.png';
+  static const String MARKER =
+      'packages/stories_editor/assets/icons/marker.png';
   static const String NEON = 'packages/stories_editor/assets/icons/neon.png';
   static const String PEN = 'packages/stories_editor/assets/icons/pen.png';
-  static const String PHOTO_FILTER = 'packages/stories_editor/assets/icons/photo_filter.png';
-  static const String PICKCOLOR = 'packages/stories_editor/assets/icons/pickColor.png';
-  static const String RETURN = 'packages/stories_editor/assets/icons/return.png';
-  static const String STICKERS = 'packages/stories_editor/assets/icons/stickers.png';
+  static const String PHOTO_FILTER =
+      'packages/stories_editor/assets/icons/photo_filter.png';
+  static const String PICKCOLOR =
+      'packages/stories_editor/assets/icons/pickColor.png';
+  static const String RETURN =
+      'packages/stories_editor/assets/icons/return.png';
+  static const String STICKERS =
+      'packages/stories_editor/assets/icons/stickers.png';
   static const String TEXT = 'packages/stories_editor/assets/icons/text.png';
   static const String TRASH = 'packages/stories_editor/assets/icons/trash.png';
-  static const String VIDEO_TRIM = 'packages/stories_editor/assets/icons/video_trim.png';
-  static const String INSTAGRAM_LOGO = 'packages/stories_editor/assets/images/instagram_logo.png';
+  static const String VIDEO_TRIM =
+      'packages/stories_editor/assets/icons/video_trim.png';
+  static const String INSTAGRAM_LOGO =
+      'packages/stories_editor/assets/images/instagram_logo.png';
 }
