@@ -21,20 +21,20 @@ class DraggableWidget extends StatelessWidget {
   final Function(PointerMoveEvent)? onPointerMove;
   final BuildContext context;
   const DraggableWidget({
-    Key? key,
+    super.key,
     required this.context,
     required this.draggableWidget,
     this.onPointerDown,
     this.onPointerUp,
     this.onPointerMove,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
     final ScreenUtil screenUtil = ScreenUtil();
-    var _colorProvider =
+    var colorProvider =
         Provider.of<GradientNotifier>(this.context, listen: false);
-    var _controlProvider =
+    var controlProvider =
         Provider.of<ControlNotifier>(this.context, listen: false);
     Widget? overlayWidget;
 
@@ -51,7 +51,7 @@ class DraggableWidget extends StatelessWidget {
               width: draggableWidget.deletePosition ? 100 : null,
               height: draggableWidget.deletePosition ? 100 : null,
               child: AnimatedOnTapButton(
-                onTap: () => _onTap(context, draggableWidget, _controlProvider),
+                onTap: () => _onTap(context, draggableWidget, controlProvider),
                 child: Stack(
                   alignment: Alignment.center,
                   children: [
@@ -59,7 +59,7 @@ class DraggableWidget extends StatelessWidget {
                       child: _text(
                           background: true,
                           paintingStyle: PaintingStyle.fill,
-                          controlNotifier: _controlProvider),
+                          controlNotifier: controlProvider),
                     ),
                     IgnorePointer(
                       ignoring: true,
@@ -67,7 +67,7 @@ class DraggableWidget extends StatelessWidget {
                         child: _text(
                             background: true,
                             paintingStyle: PaintingStyle.stroke,
-                            controlNotifier: _controlProvider),
+                            controlNotifier: controlProvider),
                       ),
                     ),
                     Padding(
@@ -77,7 +77,7 @@ class DraggableWidget extends StatelessWidget {
                           Center(
                             child: _text(
                                 paintingStyle: PaintingStyle.fill,
-                                controlNotifier: _controlProvider),
+                                controlNotifier: controlProvider),
                           ),
                         ],
                       ),
@@ -92,14 +92,14 @@ class DraggableWidget extends StatelessWidget {
 
       /// image [file_image_gb.dart]
       case ItemType.image:
-        if (_controlProvider.mediaPath.isNotEmpty) {
+        if (controlProvider.mediaPath.isNotEmpty) {
           overlayWidget = SizedBox(
             width: screenUtil.screenWidth - 144.w,
             child: FileImageBG(
-              filePath: File(_controlProvider.mediaPath),
+              filePath: File(controlProvider.mediaPath),
               generatedGradient: (color1, color2) {
-                _colorProvider.color1 = color1;
-                _colorProvider.color2 = color2;
+                colorProvider.color1 = color1;
+                colorProvider.color2 = color2;
               },
             ),
           );
@@ -236,32 +236,32 @@ class DraggableWidget extends StatelessWidget {
     if (draggableWidget.type == ItemType.text) {
       scale = 0.4;
       return scale;
-    } 
+    }
   }
 
   /// onTap text
   void _onTap(BuildContext context, EditableItem item,
       ControlNotifier controlNotifier) {
-    var _editorProvider =
+    var editorProvider =
         Provider.of<TextEditingNotifier>(context, listen: false);
-    var _itemProvider =
+    var itemProvider =
         Provider.of<DraggableWidgetNotifier>(context, listen: false);
 
     /// load text attributes
-    _editorProvider.textController.text = item.text.trim();
-    _editorProvider.text = item.text.trim();
-    _editorProvider.fontFamilyIndex = item.fontFamily;
-    _editorProvider.textSize = item.fontSize;
-    _editorProvider.backGroundColor = item.backGroundColor;
-    _editorProvider.textAlign = item.textAlign;
-    _editorProvider.textColor =
+    editorProvider.textController.text = item.text.trim();
+    editorProvider.text = item.text.trim();
+    editorProvider.fontFamilyIndex = item.fontFamily;
+    editorProvider.textSize = item.fontSize;
+    editorProvider.backGroundColor = item.backGroundColor;
+    editorProvider.textAlign = item.textAlign;
+    editorProvider.textColor =
         controlNotifier.colorList!.indexOf(item.textColor);
-    _editorProvider.animationType = item.animationType;
-    _editorProvider.textList = item.textList;
-    _editorProvider.fontAnimationIndex = item.fontAnimationIndex;
-    _itemProvider.draggableWidget
-        .removeAt(_itemProvider.draggableWidget.indexOf(item));
-    _editorProvider.fontFamilyController = PageController(
+    editorProvider.animationType = item.animationType;
+    editorProvider.textList = item.textList;
+    editorProvider.fontAnimationIndex = item.fontAnimationIndex;
+    itemProvider.draggableWidget
+        .removeAt(itemProvider.draggableWidget.indexOf(item));
+    editorProvider.fontFamilyController = PageController(
       initialPage: item.fontFamily,
       viewportFraction: .1,
     );
